add api calls

+++ code/go/0chain.net/smartcontract/storagesc/stakepool.go
@@ -255,7 +255,7 @@ func (sp *stakePool) slash(
 		if err != nil {
 			return 0, err
 		}
-		edbSlash.DelegatePenalties[id] = int64(dpSlash)
+		edbSlash.DelegatePenalties[id] = dpSlash
 	}
 	// todo we should slash from stake pools not rewards. 0chain issue 1495
 	if err := edbSlash.Emit(event.TagStakePoolReward, balances); err != nil {
